fix: 패키지 형식 구조 통일
