+++ src/main/java/com/mocamp/mocamp_backend/controller/RoomSocketController.java
@@ -2,7 +2,8 @@ package com.mocamp.mocamp_backend.controller;
 
 import com.mocamp.mocamp_backend.dto.goal.GoalCompleteUpdateRequest;
 import com.mocamp.mocamp_backend.dto.goal.GoalListRequest;
-import com.mocamp.mocamp_backend.service.goal.GoalService;
+import com.mocamp.mocamp_backend.service.goal.GoalHttpService;
+import com.mocamp.mocamp_backend.service.goal.GoalSocketService;
 import com.mocamp.mocamp_backend.service.room.RoomSocketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
@@ -15,7 +16,8 @@ import org.springframework.stereotype.Controller;
 @RequiredArgsConstructor
 public class RoomSocketController {
     private final RoomSocketService roomSocketService;
-    private final GoalService goalService;
+    private final GoalHttpService goalHttpService;
+    private final GoalSocketService goalSocketService;
 
     @MessageMapping("/video/{roomId}")
     public void streamVideo(@Header("Authorization") String token,
@@ -37,12 +39,12 @@ public class RoomSocketController {
 
     @MessageMapping("/data/goal/manage/{roomId}")
     public void manageGoal(@Payload GoalListRequest goalListRequest, @DestinationVariable("roomId") Long roomId) {
-        goalService.manageGoal(goalListRequest, roomId);
+        goalSocketService.manageGoal(goalListRequest, roomId);
     }
 
     @MessageMapping("/data/goal/complete/{roomId}")
     public void pressGoal(@Payload GoalCompleteUpdateRequest goalCompleteUpdateRequest, @DestinationVariable("roomId") Long roomId) {
-        goalService.pressGoal(goalCompleteUpdateRequest, roomId);
+        goalSocketService.pressGoal(goalCompleteUpdateRequest, roomId);
     }
 
 }
